+++ src/Providers/Provider.php
@@ -52,7 +52,10 @@ abstract class Provider extends ServiceProvider
     {
         $this->registerListeners();
         $this->mapModelMorphAliases();
-        $this->mapRoutes();
+
+        if (! $this->app->routesAreCached()) {
+            $this->mapRoutes();
+        }
     }
 
     /**
@@ -84,8 +87,6 @@ abstract class Provider extends ServiceProvider
      */
     protected function mapRoutes() : void
     {
-        if($this->app->routesAreCached() === true) return;
-
         foreach ($this->routers as $router) {
             $this->app->make($router);
         }