+++ src/Providers/ModuleProvider.php
@@ -185,6 +185,6 @@ class ModuleProvider extends Provider
      */
     protected function getLowercasePackageName() : string
     {
-        return mb_strtolower($this->getPackageName());
+        return mb_strtolower(str_slug(snake_case($this->getPackageName())));
     }
 }