+++ CHANGELOG.md
@@ -6,9 +6,12 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## Unreleased
 
+## 3.0.0 (2019-01-21)
+
 ### Added
 
 - Prevent registering routes when they are already cached
+- Prevent merging configuration when its already cached
 - Automatically register additional service providers
 
 ### Changed
@@ -20,6 +23,12 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Correctly tagged laravel version requirements
 - Correctly cached Travis Composer cache to speed up testing
 - Fixed lowercase package and config name
+
+## 1.1.2 (2018-10-22)
+
+### Fixed
+
+- Fixed autoload test
 
 ## 1.1.1 (2018-10-22)